test(e2e): migrate Name Lookup Snap smoke to Appium
